+++ lib/app.dart
@@ -1,9 +1,5 @@
-import 'package:flutter/material.dart';
-import 'package:restaurantour/core/inject.dart';
-import 'package:restaurantour/core/routes.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:restaurantour/core/theme.dart';
+import 'package:restaurantour/core/core.dart';
+import 'package:restaurantour/theme/theme.dart';
 
 class App extends StatelessWidget {
   const App({super.key});
