+++ lib/components/rt_shimmer_loading.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/material.dart';
-import 'package:restaurantour/core/rt_colors.dart';
 import 'package:shimmer/shimmer.dart';
+import 'package:restaurantour/core/core.dart';
+import 'package:restaurantour/theme/theme.dart';
 
 class RTShimmerLoading extends StatelessWidget {
   const RTShimmerLoading({super.key});
