+++ lib/core/core.dart
@@ -1,0 +1,7 @@
+export 'package:flutter_svg/svg.dart';
+export 'package:flutter/material.dart';
+export 'package:provider/provider.dart';
+export 'package:restaurantour/core/inject.dart';
+export 'package:restaurantour/core/routes.dart';
+export 'package:flutter_localizations/flutter_localizations.dart';
+export 'package:flutter_gen/gen_l10n/app_localizations.dart';
