+++ lib/theme/theme.dart
@@ -1,0 +1,3 @@
+export 'package:restaurantour/theme/rt_colors.dart';
+export 'package:restaurantour/theme/rt_text_style.dart';
+export 'package:restaurantour/theme/rt_theme.dart';
