+++ test/widget_for_tests.dart
@@ -1,7 +1,5 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:restaurantour/core/theme.dart';
+import 'package:restaurantour/core/core.dart';
+import 'package:restaurantour/theme/theme.dart';
 
 Widget widgetBuilder(Widget child) {
   return MaterialApp(
